+++ changelog.md
@@ -3,6 +3,10 @@
 #### Added
 - Mapper for Notification repository (database)
 - Mapper for Notification controller (HTTP)
+- Cancel notification useCase and tests
+- countRecipientNotifications useCase and tests
+- getRecipientNotification useCase and tests
+- Notification factory
 
 ### [0.0.3] - 2022-12-13